requirements.txt
torch==1.4
syft==0.2.9
pyjwt==1.7.1

matplotlib
gevent
gevent-websocket
pymysql
flask
flask_cors
flask_login
flask_lambda
flask_migrate
flask_sockets
flask_executor
flask_sqlalchemy
sqlalchemy_mixins
sqlalchemy_utils
sqlalchemy-aurora-data-api
gunicorn
requests
requests-toolbelt
psycopg2-binary
redis
wsaccel
papermill
bcrypt

aiounittest
black
coverage
pre-commit
pytest
isort
docformatter
flake8
flake8-comprehensions
jupyter
flask-lambda-python36

-e PyGrid
